python/requirements.txt
clearblade-cloud-iot
pyfiglet